+++ package/thunderstore/MyLittleUI/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 1.1.28
+* new option to remove last uncooked item from Cooking station and Oven (Cooking skill will be undone)
+* hold E to add ore/fuel to smelters faster than vanilla
+
 # 1.1.27
 * fixed occasional warning spam
 * font size of recipe description made configurable